requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
langchain
langchain-community==0.2.12
langchain-openai==0.1.23
pydantic==2.8.2
